Refactor logging system

+++ src/Platform/Easy.Platform/Application/MessageBus/InboxPattern/PlatformInboxBusMessageCleanerHostedService.cs
@@ -1,7 +1,6 @@
 using Easy.Platform.Application.Context;
 using Easy.Platform.Common.Extensions;
 using Easy.Platform.Common.Hosting;
-using Easy.Platform.Common.Timing;
 using Easy.Platform.Common.Utils;
 using Easy.Platform.Domain.UnitOfWork;
 using Microsoft.Extensions.DependencyInjection;
@@ -50,7 +49,7 @@ public class PlatformInboxBusMessageCleanerHostedService : PlatformIntervalProce
         {
             // WHY: Retry in case of the db is not started, initiated or restarting
             await Util.TaskRunner.WaitRetryThrowFinalExceptionAsync(
-                () => CleanInboxEventBusMessage(cancellationToken),
+                async () => await CleanInboxEventBusMessage(cancellationToken),
                 retryAttempt => 10.Seconds(),
                 retryCount: ProcessClearMessageRetryCount(),
                 onRetry: (ex, timeSpan, currentRetry, ctx) =>
@@ -106,36 +105,49 @@ public class PlatformInboxBusMessageCleanerHostedService : PlatformIntervalProce
 
     protected async Task CleanInboxEventBusMessage(CancellationToken cancellationToken)
     {
-        await ServiceProvider.ExecuteInjectScopedAsync(
-            async (IUnitOfWorkManager uowManager, IPlatformInboxBusMessageRepository inboxEventBusMessageRepo) =>
-            {
-                using (var uow = uowManager!.Begin())
+        var toCleanMessageCount = await ServiceProvider.GetRequiredService<IPlatformInboxBusMessageRepository>()
+            .CountAsync(
+                PlatformInboxBusMessage.ToCleanInboxEventBusMessagesExpr(DeleteProcessedMessageInSeconds(), DeleteExpiredFailedMessageInSeconds()),
+                cancellationToken);
+
+        if (toCleanMessageCount > 0)
+        {
+            await Util.Pager.ExecuteScrollingPagingAsync(
+                async () =>
                 {
-                    var expiredMessages = await inboxEventBusMessageRepo.GetAllAsync(
-                        queryBuilder: query => query
-                            .Where(
-                                p => (p.LastConsumeDate <= Clock.UtcNow.AddSeconds(-DeleteProcessedMessageInSeconds()) &&
-                                      p.ConsumeStatus == PlatformInboxBusMessage.ConsumeStatuses.Processed) ||
-                                     (p.LastConsumeDate <= Clock.UtcNow.AddSeconds(-DeleteExpiredFailedMessageInSeconds()) &&
-                                      p.ConsumeStatus == PlatformInboxBusMessage.ConsumeStatuses.Failed))
-                            .OrderBy(p => p.LastConsumeDate)
-                            .Take(NumberOfDeleteMessagesBatch()),
-                        cancellationToken);
+                    return await ServiceProvider.ExecuteInjectScopedAsync<List<PlatformInboxBusMessage>>(
+                        async (IUnitOfWorkManager uowManager, IPlatformInboxBusMessageRepository inboxEventBusMessageRepo) =>
+                        {
+                            using (var uow = uowManager!.Begin())
+                            {
+                                var expiredMessages = await inboxEventBusMessageRepo.GetAllAsync(
+                                    queryBuilder: query => query
+                                        .Where(
+                                            PlatformInboxBusMessage.ToCleanInboxEventBusMessagesExpr(
+                                                DeleteProcessedMessageInSeconds(),
+                                                DeleteExpiredFailedMessageInSeconds()))
+                                        .OrderBy(p => p.LastConsumeDate)
+                                        .Take(NumberOfDeleteMessagesBatch()),
+                                    cancellationToken);
 
-                    if (expiredMessages.Count > 0)
-                    {
-                        await inboxEventBusMessageRepo.DeleteManyAsync(
-                            expiredMessages,
-                            dismissSendEvent: true,
-                            cancellationToken);
+                                if (expiredMessages.Count > 0)
+                                {
+                                    await inboxEventBusMessageRepo.DeleteManyAsync(
+                                        expiredMessages,
+                                        dismissSendEvent: true,
+                                        cancellationToken);
 
-                        await uow.CompleteAsync(cancellationToken);
+                                    await uow.CompleteAsync(cancellationToken);
+                                }
 
-                        Logger.LogInformation(
-                            message:
-                            $"CleanInboxEventBusMessage success. Number of deleted messages: {expiredMessages.Count}");
-                    }
-                }
-            });
+                                return expiredMessages;
+                            }
+                        });
+                });
+
+            Logger.LogInformation(
+                message:
+                $"CleanInboxEventBusMessage success. Number of deleted messages: {toCleanMessageCount}");
+        }
     }
 }
